Fix a dastardly math bug in `Malloc`.
(Un)Fortunately, the statements

    -    n = (n + 8) & ~8;
    +    n = (n + 7) & ~7;

have very similar effects when `n` is already a multiple of 8!

Also, eliminate a use of `Malloc` that could just be done on the stack,
when we know we're doing Sudoku and so there will be exactly 81 rows
in each solution.

We're now running at about "129500/sec" on my MacBook.