"Fill uncontested cells" before outputting each solution.
For example, no longer output a solution containing a corner like this:

    . . . .
    ........
    ..22.....
    ..222...
    ..2.22...

when the "hole" could have been filled in like this:

    . . . .
    ........
    ..22.....
    ..222...
    ..2222...

I'm not sure this is a win, especially for the non-Bernie version,
because it tends to place queens that then need to be "un-placed"
in order to free up cells for queens of the biggest-army color.
But it saves me from having to hand-polish these solutions.